% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cb_save}
\alias{cb_save}
\title{Custom save function}
\usage{
cb_save(gg, name, w = 10, h = 4, location = "plots")
}
\arguments{
\item{gg}{a ggplot object/chart to save}

\item{name}{the name of the png}

\item{w}{width of the saved picture}

\item{h}{height of the saved picture}

\item{location}{folder in the root directory in which to save plot}
}
\value{
nothing
}
\description{
Save ggplot objects as png with set height and width
}
