% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_coordinates}
\alias{convert_to_coordinates}
\title{Convert city and region into coordinates}
\usage{
convert_to_coordinates(city, region = "")
}
\arguments{
\item{city}{name of a city}

\item{region}{name of the region e.g. england, northern ireland}
}
\value{
coordinates
}
\description{
Convert city and region into coordinates
}
