% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_and_login_information.R
\name{get_account_activity}
\alias{get_account_activity}
\title{Get account activity}
\usage{
get_account_activity(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of account activity from the last 3 years
}
\description{
Get account activity
}
