% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads_and_businesses.R
\name{get_ad_interests}
\alias{get_ad_interests}
\title{Get ads interests}
\usage{
get_ad_interests(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
character vector of all the users ad interests
}
\description{
Get ads interests
}
