% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_you.R
\name{get_address_book}
\alias{get_address_book}
\title{Get address book}
\usage{
get_address_book(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of the users mobile phone contacts
}
\description{
Get address book
}
