% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_and_login_information.R
\name{get_admin_records}
\alias{get_admin_records}
\title{Get administrative records}
\usage{
get_admin_records(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of admin records
}
\description{
Get administrative records
}
