% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads_and_businesses.R
\name{get_advert_interactions}
\alias{get_advert_interactions}
\title{Get advertisers the user has interacted with (in the last month?)}
\usage{
get_advert_interactions(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of advertisers and the users interaction with their advert
}
\description{
Get advertisers the user has interacted with (in the last month?)
}
