% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads_and_businesses.R
\name{get_advertisers_with_contact_info}
\alias{get_advertisers_with_contact_info}
\title{Get advertisers who have your contact info}
\usage{
get_advertisers_with_contact_info(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
character vector of advertisers who have uploaded a contact list containing
the users information
}
\description{
Get advertisers who have your contact info
}
