% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_and_login_information.R
\name{get_auth_logins}
\alias{get_auth_logins}
\title{Get authorised logins}
\usage{
get_auth_logins(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of authorised logins
}
\description{
Get authorised logins
}
