% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{get_comments}
\alias{get_comments}
\title{Get comments}
\usage{
get_comments(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of comments
}
\description{
Get comments
}
