% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_event_invitations}
\alias{get_event_invitations}
\title{Get event invitations}
\usage{
get_event_invitations(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of event invitations
}
\description{
Get event invitations
}
