% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_event_responses}
\alias{get_event_responses}
\title{Get event responses}
\usage{
get_event_responses(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of event responses
}
\description{
Get event responses
}
