% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_you.R
\name{get_events_visited}
\alias{get_events_visited}
\title{Get Events visited}
\usage{
get_events_visited(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of events visited with date/time and name
}
\description{
Get Events visited
}
