% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likes_and_reactions.R
\name{get_external_likes}
\alias{get_external_likes}
\title{Get likes from external sites}
\usage{
get_external_likes(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of likes from external sites
}
\description{
Get likes from external sites
}
