% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_you.R
\name{get_face_recognition_info}
\alias{get_face_recognition_info}
\title{Get facial recognition info}
\usage{
get_face_recognition_info(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
the number of examples of the users face they have
}
\description{
Get facial recognition info
}
