% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_you.R
\name{get_friend_peer_group}
\alias{get_friend_peer_group}
\title{Get peer group from Facebook data file}
\usage{
get_friend_peer_group(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
a character string denoting the 'friend peer group' assigned by Facebook
}
\description{
Get peer group from Facebook data file
}
