% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_activity.R
\name{get_fundraisers_donated_to}
\alias{get_fundraisers_donated_to}
\title{Get fundraisers donated to}
\usage{
get_fundraisers_donated_to(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of fundraisers donated to
}
\description{
Get fundraisers donated to
}
