% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_you.R
\name{get_groups_visited}
\alias{get_groups_visited}
\title{Get groups visited}
\usage{
get_groups_visited(folder = "data")
}
\description{
@param folder the name of the data folder (in the project root directory)
}
\details{
@return data frame of groups visited by the user and date/time of last visit
}
