% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps_and_websites.R
\name{get_installed_apps}
\alias{get_installed_apps}
\title{Get installed apps (apps linked to FB account)}
\usage{
get_installed_apps(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of all installed apps and when they were installed
}
\description{
Get installed apps (apps linked to FB account)
}
