% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marketplace.R
\name{get_items_sold}
\alias{get_items_sold}
\title{Get items sold}
\usage{
get_items_sold(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
dataframe of items sold
}
\description{
Get items sold
}
