% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_last_location}
\alias{get_last_location}
\title{Get last location}
\usage{
get_last_location(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of time and coordinates of last known location
}
\description{
Get last location
}
