% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likes_and_reactions.R
\name{get_liked_pages}
\alias{get_liked_pages}
\title{Get 'liked' pages}
\usage{
get_liked_pages(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of liked pages
}
\description{
Get 'liked' pages
}
