% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_location_history}
\alias{get_location_history}
\title{Get location history}
\usage{
get_location_history(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
a data frame of user location history
}
\description{
Get location history
}
