% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_and_login_information.R
\name{get_login_protection_data}
\alias{get_login_protection_data}
\title{Get login protection data}
\usage{
get_login_protection_data(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of login protection data
}
\description{
Get login protection data
}
