% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/security_and_login_information.R
\name{get_logins_and_outs}
\alias{get_logins_and_outs}
\title{Get logins and logouts}
\usage{
get_logins_and_outs(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of logins and logouts
}
\description{
Get logins and logouts
}
