% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_you.R
\name{get_marketplace_interactions}
\alias{get_marketplace_interactions}
\title{Get Marketplace interactions}
\usage{
get_marketplace_interactions(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
list of data frames of marketplace interactions
}
\description{
Get Marketplace interactions
}
