% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{get_messages}
\alias{get_messages}
\title{Get messages with an individual}
\usage{
get_messages(folder = "data", participant)
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}

\item{participant}{the name of someone who you have sent or received messages from
(will search for this name in the data/messages/inbox folder)}
}
\value{
a character string denoting the 'friend peer group' assigned by Facebook
}
\description{
Get messages with an individual
}
