% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{get_my_events}
\alias{get_my_events}
\title{Get events created by user}
\usage{
get_my_events(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of users created events
}
\description{
Get events created by user
}
