% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads_and_businesses.R
\name{get_off_facebook_activity}
\alias{get_off_facebook_activity}
\title{Get off-Facebook activity}
\usage{
get_off_facebook_activity(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of off-facebook activity
}
\description{
Get off-Facebook activity
}
