% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{get_posts}
\alias{get_posts}
\title{Get users posts}
\usage{
get_posts(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of the user's posts on facebook
}
\description{
Get users posts
}
