% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps_and_websites.R
\name{get_posts_from_external}
\alias{get_posts_from_external}
\title{Get facebook posts from external websites}
\usage{
get_posts_from_external(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of posts made to users profile by external apps/websites
}
\description{
Get facebook posts from external websites
}
