% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_primary_location}
\alias{get_primary_location}
\title{Get primary location}
\usage{
get_primary_location(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame with one observation: the users primary known location
}
\description{
Get primary location
}
