% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_information.R
\name{get_profile_information}
\alias{get_profile_information}
\title{Get profile information}
\usage{
get_profile_information(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
a list of data frames containing profile information
}
\description{
Get profile information
}
