% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_information.R
\name{get_profile_update_history}
\alias{get_profile_update_history}
\title{Get profile update history}
\usage{
get_profile_update_history(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
a list of data frames containing the profile update history
}
\description{
Get profile update history
}
