% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likes_and_reactions.R
\name{get_reactions}
\alias{get_reactions}
\title{Get reactions}
\usage{
get_reactions(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of all reactions to posts and comments
}
\description{
Get reactions
}
