% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_you.R
\name{get_recent_ads}
\alias{get_recent_ads}
\title{Get Ads viewed in the last week}
\usage{
get_recent_ads(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
data frame of ads recently viewed on Facebook
}
\description{
Get Ads viewed in the last week
}
