% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_used_for_recommendations.R
\name{get_recommended_topics}
\alias{get_recommended_topics}
\title{Get information used for recommendations}
\usage{
get_recommended_topics(type, folder = "data")
}
\arguments{
\item{type}{any of 'watch', 'news' or 'feed'. Will return recommended topics for: Facebook Watch;
news stories; and the users newsfeed, respectively.}

\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
a character vector of topics used for recommendations
}
\description{
Get information used for recommendations
}
