% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_you.R
\name{get_secret_conversations}
\alias{get_secret_conversations}
\title{Get any secret conversations}
\usage{
get_secret_conversations(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
vector of logicals indicating whether the user has sent or received any secret conversations
}
\description{
Get any secret conversations
}
