% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{get_timezone}
\alias{get_timezone}
\title{Get timezone}
\usage{
get_timezone(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
the users timezone
}
\description{
Get timezone
}
