% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_you.R
\name{get_videos_watched}
\alias{get_videos_watched}
\title{Get information about videos watched}
\usage{
get_videos_watched(folder = "data")
}
\arguments{
\item{folder}{the name of the data folder (in the project root directory)}
}
\value{
list of data frames giving information about videos watched on Facebook Watch
}
\description{
Get information about videos watched
}
