% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_temporal}
\alias{gg_temporal}
\title{Plot temporal chart with periods of life}
\usage{
gg_temporal(data, aes, label_height = 60, type = "date")
}
\arguments{
\item{data}{ggplot-compatible dataframe}

\item{aes}{ggplot aesthetics}

\item{label_height}{the value of y at which to put the period of life annotations}

\item{type}{either "date" or "numeric" depending on the class of the x axis}
}
\value{
a base ggplot object
}
\description{
Plot temporal chart with periods of life
}
